+++ NEWS.md
@@ -1,4 +1,4 @@
-# flattabler (development version)
+# flattabler 2.1.0
 * Add `keep_col_names` and `na_rm` parameters to `unpivot` function.
 * Read the indicated sheets from a folder of Excel files in `read_excel_folder()`.